+++ src/query/react/index.ts
@@ -1,3 +1,4 @@
 export * from "./useResource";
+export * from "./useSuspenseResource";
 export * from "./useCommand";
 export * from "./ReactHooksPlugin";